Support to install pointgrey sdk

+++ jenkins-scripts/docker/lib/subt-compilation-base.bash
@@ -16,19 +16,31 @@ DOCKER_JOB_NAME="subt_ci"
 . ${SCRIPT_DIR}/lib/_gazebo_utils.sh
 . ${SCRIPT_DIR}/lib/_subt_utils.sh
 
+# Needs pointgrey SDK (manual installation) and downloader is broken from ros-drivers.
+# manual installation from our nodes
+if [[ -d /var/lib/jenkins/.pointgrey_sdk ]]; then
+    echo 'The node has not been setup for pointgrey sdk'
+    exit 1
+fi
+
+mkdir -p "${WORKSPACE}/pointgrey/"
+for f in libflycapture-2.13.3.31_amd64.deb libflycapture-2.13.3.31_amd64-dev.deb; do
+  cp "/var/lib/jenkins/.pointgrey_sdk/$f" "${WORKSPACE}/pointgrey/"
+done
+
 # husky and jackal descriptions are used
 # Models are used in tests
 export ROS_WS_PREBUILD_HOOK="""
+# Install pointgrey sdk
+dpkg -i ${WORKSPACE}/pointgrey/libflycapture-2.13.3.31_amd64.deb
+dpkg -i ${WORKSPACE}/pointgrey/libflycapture-2.13.3.31_amd64-dev.deb
 rm -fr ${WORKSPACE}/subt/husky_description ${WORKSPACE}/subt/pointgrey_camera_driver ${WORKSPACE}/subt/jackal_description
 git clone --depth 1 https://github.com/husky/husky.git /tmp/huksy > /dev/null
 git clone --depth 1 https://github.com/jackal/jackal.git /tmp/jackal > /dev/null
+git clone --depth 1 https://github.com/ros-drivers/pointgrey_camera_driver /tmp/pointgrey_camera_driver > /dev/null
 mv /tmp/huksy/husky_description ${WORKSPACE}/subt/
 mv /tmp/jackal/jackal_description  ${WORKSPACE}/subt/
-# Needs pointgrey SDK (manual installation) and downloader is broken from ros-drivers.
-# git clone --depth 1 https://github.com/ros-drivers/pointgrey_camera_driver /tmp/pointgrey_camera_driver
-# mv /tmp/pointgrey_camera_driver ${WORKSPACE}/subt/
-# wget https://raw.githubusercontent.com/ros-drivers/pointgrey_camera_driver/master/pointgrey_camera_driver/cmake/download_flycap
-# python download_flycap
+mv /tmp/pointgrey_camera_driver ${WORKSPACE}/subt/
 ${GAZEBO_MODEL_INSTALLATION}
 """
